requirements.txt
# Python Backend Dependencies
Flask==2.3.3
Werkzeug==2.3.7
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
flask_pymongo==2.3.0
pymongo[srv]==4.6.1
certifi==2023.7.22
python-dotenv==0.19.0
bcrypt==3.2.0
requests==2.26.0
gunicorn==20.1.0 